Merged revisions 117295 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r117295 | rizzo | 2008-05-20 19:48:43 +0200 (Tue, 20 May 2008) | 10 lines

Document the possible presence of multiple variables with the
same name in http queries, which might confuse the manager.

Replace calls to ast_uri_decode() with a local function that also
replaces '+' with ' ', as this is the normal encoding for
spaces in http requests.
This allows passing cli commands to the manager through the
http interface.


........