Fix a race condition that can cause an IAX2 call to end up in a storm of INVAL and VNAK frames being sent.
The recent security fix was to make chan_iax2 do more strict checking
of incoming full frames.  Part of this fix was to ensure that all full frames
that were supposed to include the destination call number, did include it.
So, the 2nd frame received for a call is expected to include the destination
call number.  However, this is where a problem can occur.

The IAX2 channel driver schedules a PING and LAGRQ to be sent right after an
IAX2 exchange begins.  One of these frames (usually the LAGRQ) may be sent as
the second frame before the first frame is received from the other end.  If that
happens, then the frame will be rejected as invalid now, and for every time in
the future as it gets retransmitted.

I have made a minor change to the callbacks that send LAGRQ and PING to ensure
that they do not get sent unless the peer's call number is already known.