Creating a branch to solve the issue of not being able to log attended transfers in the queue_log.
I believe that by using a datastore on the queue member's channel, an attended transfer
can properly be logged during the "fixup" portion of a masquerade. The potentially tricky
part is being certain that such events are not logged on every masquerade. The result of
that could be that those who use Local channels as queue members see TRANSFER for every
call taken.

The work on this should not be too difficult, but I'm creating a developer branch to be
on the safe side