Blocked revisions 125133 via svnmerge
........
r125133 | mmichelson | 2008-06-25 17:25:20 -0500 (Wed, 25 Jun 2008) | 11 lines

Fix a bug in the rwlock tracking. ast_rwlock_unlock did not take into
account that multiple threads could hold the same rdlock at the same time.
As such, it expected that when a thread released a lock that it must have
been the last to acquire the lock as well. Erroneous error messages would
be sent to the console stating that a thread was attempting to unlock a lock
it did not own.

Now all threads are examined to be sure that the message is only printed 
when it is supposed to be printed.


........