Changed instances of make_gsm_file to make_file.
Why? Well, for one thing they accomplish the 
same task except that make_gsm_file for some 
silly reason had a call to mkdir() in it...and 
it more or less ignored the return value of 
the mkdir() call.

Also, make_gsm_file had absolutely _nothing_ 
at all to do with gsm whatsoever so its name 
was just plain stupid and misleading.

Furthermore, the mode in the mkdir() call was
set to 01777. This seems overly silly to me.

Oh, and for some reason, an unused "imapuser"
string was passed to make_gsm_file.

Good riddance!
</rant>