Progress. I had been experiencing crashes and the reason was due to a perfectly valid assumption I had regarding the vm_state retrieval functions.
For get_vm_state_by_imapuser, the "interactive" parameter may
be 0, 1, or 2. 2 means to not try to match based on the saved
vm_state's interactive field. For get_vm_state_by_mailbox, the
interactive parameter always is attempted to be matched with
the saved vm_state's interactive field. This means that passing
a 2 will never ever match. Stupid stupid inconsistency.