Fix the "delete" option in voicemail.conf by doing the following.
1. Change IMAP's DELETE macro to flag the message as 
"Deleted." This makes it more parallel with what ODBC
does for its DELETE callback.
2. Remove the IMAP_DELETE macro since it was only used
in one place and should be taken care of by the DELETE
macro instead. (Speaking of IMAP_DELETE, if you're going 
to have a macro which is only used by IMAP and only used 
in one place, why have the macro take four arguments when
the function called within the macro only takes two?)