Merged revisions 136544 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r136544 | sruffell | 2008-08-07 13:14:17 -0500 (Thu, 07 Aug 2008) | 11 lines

Updated codec_dahdi to use the new transcoder interface in the first DAHDI
release.  Codec dahdi no longer functions with the transcoder interface in
zaptel at this time (which the last zaptel release was 1.4.11).

NOTE: Still needs an update to the configure script to make sure that
codec_dahdi is only built if the new transcoder interface is present in the
drivers.

(Issue: DAHDI-42)


........