Merged revisions 136715,136720,136722,136746,136751 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r136715 | mmichelson | 2008-08-07 18:25:50 -0400 (Thu, 07 Aug 2008) | 18 lines

Merging the imap_consistency_trunk branch to
trunk.

For an explanation of what "imap_consistency" is,
please see svn revision 134223 to the 1.4 branch.

Coincidentally, this also fixes a recent bug report
regarding the inability to save messages to the new
folder when using IMAP storage since they will would
be flagged as "seen" and not be recognized as new
messages.


(closes issue #13234)
Reported by: jaroth



................
r136720 | mmichelson | 2008-08-07 19:38:08 -0400 (Thu, 07 Aug 2008) | 6 lines

Fix build for non-IMAP storage and get rid of some
debug messages.
Thanks to eliel for alerting me.
No thanks to buildbot.


................
r136722 | mmichelson | 2008-08-07 19:39:50 -0400 (Thu, 07 Aug 2008) | 3 lines

Remove one last batch of debug messages


................
r136746 | murf | 2008-08-07 20:48:35 -0400 (Thu, 07 Aug 2008) | 40 lines

Merged revisions 136726 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r136726 | murf | 2008-08-07 18:15:34 -0600 (Thu, 07 Aug 2008) | 32 lines


(closes issue #13236)
Reported by: korihor

Wow, this one was a challenge!

I regrouped and ran a new strategy for
setting the ~~MACRO~~ value; I set it once
per extension, up near the top. It is only
set if there is a switch in the extension.

So, I had to put in a chunk of code to detect
a switch in the pval tree.

I moved the code to insert the set of ~~exten~~
up to the beginning of the gen_prios routine, 
instead of down in the switch code.

I learned that I have to push the detection
of the switches down into the code, so everywhere
I create a new exten in gen_prios, I make sure
to pass onto it the values of the mother_exten
first, and the exten next.

I had to add a couple fields to the exten
struct to accomplish this, in the ael_structs.h
file. The checked field makes it so we don't
repeat the switch search if it's been done.

I also updated the regressions.


........

................
r136751 | tilghman | 2008-08-07 22:34:17 -0400 (Thu, 07 Aug 2008) | 2 lines

Removing bad properties

................