Janitor (Move from 'static const char *' to 'static char *').
Now we have syntax, description and synopsis of every application/function defined as 'char *'. This is an attempt to
move every parameter in a register_application as custom_function_register to use 'char *' instead of 'const char *' when
referencing a 'description' or a 'synopsys'.

Also the casting made in ast_register_application2() to prevent WARNINGs was removed.