Merged revisions 136819,136859 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r136819 | seanbright | 2008-08-08 14:19:33 -0400 (Fri, 08 Aug 2008) | 3 lines

Bring in the configure and makeopts jazz for -Wshadow, but don't add it to the
Makefile yet.

........
r136859 | tilghman | 2008-08-08 21:15:38 -0400 (Fri, 08 Aug 2008) | 4 lines

Update documentation as to the behavior of AGI in 1.6.0 and higher.  Also, add
an OOB message that answers the question of, if AGI no longer shuts down the
connection on hangup, how will FastAGI know when to stop processing the call?

........