Merged revisions 140052-140053,140057,140061 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
................
r140052 | twilson | 2008-08-26 10:27:46 -0500 (Tue, 26 Aug 2008) | 11 lines

Blocked revisions 140050 via svnmerge

........
r140050 | twilson | 2008-08-26 10:22:58 -0500 (Tue, 26 Aug 2008) | 4 lines

sounds/Makefile installs sounds using the "new" language directory structure, but languageprefix needs to be set = yes for sounds in subdirectories (digits/1, etc.) to play as the correct language.

Fix the generation of asterisk.conf to include languageprefix=yes

........

................
r140053 | russell | 2008-08-26 10:29:25 -0500 (Tue, 26 Aug 2008) | 23 lines

Merged revisions 140051 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r140051 | russell | 2008-08-26 10:27:23 -0500 (Tue, 26 Aug 2008) | 15 lines

Fix a race condition with the IAX scheduler thread.  A lock and condition are
used here to allow newly scheduled tasks to wake up the scheduler just in case
the new task needs to run sooner than the current wakeup time when the thread
is sleeping.  However, there was a race condition such that a newly scheduled
task would not properly wake up the scheduler or affect the wake up period.
The order of execution would have been:

  1) Scheduler thread determines wake up time of N ms.
  2) Another thread schedules a task and signals the condition, with an
     execution time of < N ms.
  3) Scheduler thread locks and goes to sleep for N ms.

By moving the sleep time determination to inside the critical section, this
possibility is avoided.

........

................
r140057 | murf | 2008-08-26 10:57:49 -0500 (Tue, 26 Aug 2008) | 21 lines

(closes issue #13366)
Reported by: erousseau

This was a reasonable enhancement request, which was
easy to implement. Since it's an enhancement, it 
could only be applied to trunk.

Basically, for accounting where "initiated" seconds
are billed for, if the microseconds field on the end
time is greater than the microseconds field for the
answer time, add one second to the billsec field.

The implementation was requested by erousseau, and
I've implemented it as requested. I've updated the
CHANGES, the cdr.conf.sample, and the .h files
accordingly, to accept and set a flag for the
corresponding new option. cdr.c adds in the extra
second based on the usec fields if the option is
set. Tested, seems to be working fine.


................
r140061 | russell | 2008-08-26 11:10:06 -0500 (Tue, 26 Aug 2008) | 14 lines

Merged revisions 140060 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r140060 | russell | 2008-08-26 11:07:58 -0500 (Tue, 26 Aug 2008) | 6 lines

Fix some bogus scheduler usage in chan_sip.  This code used the return value
of a completely unrelated function to determine whether the scheduler should
be run or not.  This would have caused the scheduler to not run in cases where
it should have.  Also, leave a note about another scheduler issue that needs
to be addressed at some point.

........

................