Progress. This code does not compile currently.
The first step was to add container-specific callbacks
so that each container operation can be customized
depending on the container used. The ao2_container_impl
struct in astobj2.h handles this.

These callbacks are now peppered throughout astobj2.c
in place of the hash table-specific code. The one function
that has not been taken care of yet is ao2_container_alloc.
It will be taken care of in the next phase, which is to
port the former hash table material to its own file.