I encountered the problem that the container-specific code could not access fields of the ao2_container struct and could not use helper functions like INTERNAL_OBJ and EXTERNAL_OBJ. I wanted to use some trickery to make those functions available to the container code but not to any other code in Asterisk.
While such trickery could probably be accomplished
through passing some pointers and documenting some
necessary semantics for the containers, I felt that
this approach is a good compromise.

With this, I am adding the file
include/asterisk/astobj2_private.h
This file contains the complete structure definition
for ao2_container, as well as the INTERNAL_OBJ
and EXTERNAL_OBJ routines. This file should only
be included by files in the astobj2 family.

Next job is to implement the hashtable allocation
function. Once that is done, I can actually start
trying to compile and test to make sure everything
still works as it is supposed to.