First draft of astobj2_skiplist.c
There are probably some kinks to work out, including
what I've placed in XXX comments. Specifically, what
needs to be done is to make skiplist_callback execute
a quick traversal to search for an item if OBJ_POINTER
is one of the search flags passed in.

Also in this commit, I have fixed some memory leaks
which would occur due to allocation failures and 
I fixed a rather grievous error in the linkedlist_link
functions.