Progress towards iterator code. The implementation for astobj2_linkedlist is a bit ugly since nearly the exact same code is duplicated twice.
The implementation for astobj2_skiplist needs work.
I tried to take the approach of using what the
ao2_iterator structure had to offer without having
to add private data (since most containers really
don't need to store any extra data in the iterator).
It appears, though, that the "version" method of
finding the next item when iterating will not
work with skiplists since items are added to the
list in any order.