Blocked revisions 141906 via svnmerge
........
r141906 | mmichelson | 2008-09-08 18:00:06 -0500 (Mon, 08 Sep 2008) | 9 lines

Optimization: The only reason we should check member
status is if the queue has a joinempty or a leavewhenempty
setting which could cause the caller to not join the queue
or exit the queue. Prior to this patch, we could potentially
traverse the entire queue's member list for no reason since
even if the members are currently not available in some way
we're going to let the caller join the queue anyway.


........